+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The tag [OUTPUT] indicates changes affecting the output files.
 ### Fixed
 
 - [OUTPUT] Fix a crash when running `metropolis` with an output directory that does not exist.
+- [DEV] Fix nodes being duplicated when a graph is deserialized.
 
 ## [0.7.0] - 2023-11-16